--- app/src/main/java/com/sopt/now/compose/data/api/dto/request/RequestSignUpDto.kt
@@ -1,28 +1,0 @@
-package com.sopt.now.compose.data.api.dto.request
-
-import com.sopt.now.compose.model.User
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RequestSignUpDto(
-    @SerialName("authenticationId")
-    val authenticationId: String,
-    @SerialName("nickname")
-    val nickname: String,
-    @SerialName("password")
-    val password: String,
-    @SerialName("phone")
-    val phone: String
-) {
-    companion object {
-        fun toSignUpDto(entity: User): RequestSignUpDto {
-            return RequestSignUpDto(
-                authenticationId = entity.id,
-                nickname = entity.nickName,
-                password = entity.password.toString(),
-                phone = entity.phone
-            )
-        }
-    }
-}
