+++ app/src/main/java/com/sopt/now/compose/data/api/dto/request/SignUpRequest.kt
@@ -1,0 +1,16 @@
+package com.sopt.now.compose.data.api.dto.request
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SignUpRequest(
+    @SerialName("authenticationId")
+    val authenticationId: String,
+    @SerialName("nickname")
+    val nickname: String,
+    @SerialName("password")
+    val password: String,
+    @SerialName("phone")
+    val phone: String
+)
